# web/openapi.json
{
  "components": {
    "schemas": {
      "AddRequest": {
        "properties": {
          "external_key": {
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": ["source_id", "external_key"],
        "type": "object"
      },
      "CatalogItemDto": {
        "description": "A catalog result, which has no local identity.",
        "properties": {
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "content_rating": {
            "type": "string"
          },
          "cover_url": {
            "type": ["string", "null"]
          },
          "description": {
            "type": ["string", "null"]
          },
          "external_key": {
            "description": "The source's own key. This, plus the source id, is how it is added.",
            "type": "string"
          },
          "manga_id": {
            "description": "Set when this series is already in the library, so the UI can offer\n\"open\" rather than \"add\".",
            "format": "uuid",
            "type": ["string", "null"]
          },
          "reading_direction": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": ["string", "null"]
          }
        },
        "required": [
          "external_key",
          "title",
          "authors",
          "artists",
          "tags",
          "status",
          "content_rating",
          "reading_direction"
        ],
        "type": "object"
      },
      "CatalogPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogItemDto"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        },
        "required": ["items"],
        "type": "object"
      },
      "ChapterDto": {
        "properties": {
          "external_key": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "language": {
            "type": ["string", "null"]
          },
          "manga_id": {
            "format": "uuid",
            "type": "string"
          },
          "number": {
            "format": "float",
            "type": ["number", "null"]
          },
          "published_at": {
            "format": "date-time",
            "type": ["string", "null"]
          },
          "title": {
            "type": ["string", "null"]
          },
          "volume": {
            "format": "float",
            "type": ["number", "null"]
          }
        },
        "required": ["id", "manga_id", "external_key"],
        "type": "object"
      },
...
      "MangaDto": {
        "properties": {
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "content_rating": {
            "type": "string"
          },
          "cover_url": {
            "type": ["string", "null"]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "external_key": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "language": {
            "type": ["string", "null"]
          },
          "reading_direction": {
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": ["string", "null"]
          }
        },
        "required": [
          "id",
          "source_id",
          "external_key",
          "title",
          "authors",
          "artists",
          "tags",
          "status",
          "content_rating",
          "reading_direction",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "Paged_MangaDto": {
        "description": "One page of results.\n\n`next_cursor` is opaque and is the only supported way to ask for the next\npage. Offsets are deliberately absent: they shift under inserts, and a\nlibrary that gains a chapter mid-scroll would skip or repeat one.",
        "properties": {
          "items": {
            "items": {
              "properties": {
                "artists": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "authors": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "content_rating": {
                  "type": "string"
                },
                "cover_url": {
                  "type": ["string", "null"]
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "description": {
                  "type": ["string", "null"]
                },
                "external_key": {
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "language": {
                  "type": ["string", "null"]
                },
                "reading_direction": {
                  "type": "string"
                },
                "source_id": {
                  "format": "uuid",
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "tags": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": {
                  "type": "string"
                },
                "updated_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "url": {
                  "type": ["string", "null"]
                }
              },
              "required": [
                "id",
                "source_id",
                "external_key",
                "title",
                "authors",
                "artists",
                "tags",
                "status",
                "content_rating",
                "reading_direction",
                "created_at",
                "updated_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        },
        "required": ["items"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "An opaque, signed, HttpOnly session cookie issued by `GET /api/v1/auth/callback`. There is no bearer token: tokens from the identity provider never reach the browser.",
        "in": "cookie",
        "name": "id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted manga library server.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "nyuka",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/manga": {
      "get": {
        "operationId": "listLibrary",
        "parameters": [
          {
            "description": "The `next_cursor` from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paged_MangaDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "`GET /api/v1/manga`",
        "tags": ["library"]
      },
      "post": {
        "description": "The body names a source and a key, not a series: the metadata comes from\nthe source, never from the client. Accepting a client's version would let\nanyone write any title into the library.",
        "operationId": "addMangaToLibrary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaDto"
                }
              }
            },
            "description": "Added, or already present"
          },
          "404": {
            "description": "No such source or series"
          },
          "502": {
            "description": "The source could not be reached"
          }
        },
        "summary": "`POST /api/v1/manga` — add a catalog entry to the library.",
        "tags": ["catalog"]
      }
    },
...
    "/sources/{id}/catalog": {
      "get": {
        "operationId": "browseCatalog",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Free-text search. Absent means the source's default listing.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPage"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such source"
          },
          "502": {
            "description": "The source could not be reached"
          }
        },
        "summary": "`GET /api/v1/sources/{id}/catalog`",
        "tags": ["catalog"]
      }
    },
    "/sources/{id}/catalog/{key}": {
      "get": {
        "operationId": "getCatalogItem",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The source's own key for the series",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such source or series"
          },
          "502": {
            "description": "The source could not be reached"
          }
        },
        "summary": "`GET /api/v1/sources/{id}/catalog/{key}`",
        "tags": ["catalog"]
      }
    },
    "/sources/{id}/catalog/{key}/chapters": {
      "get": {
        "operationId": "listCatalogChapters",
        "parameters": [
          {
            "description": "Source id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The source's own key for the series",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "No such source or series"
          },
          "502": {
            "description": "The source could not be reached"
          }
        },
        "summary": "`GET /api/v1/sources/{id}/catalog/{key}/chapters`",
        "tags": ["catalog"]
      }
    },
...
  },
...
}